impersonating, gates and policies

+++ app/Livewire/Seller/Index.php
@@ -9,11 +9,13 @@ class Index extends Component
 {
     public function render()
     {
+        $this->authorize('viewAny', Seller::class);
         return view('livewire.seller.index');
     }
 
     public function destroy(Seller $seller)
     {
+        $this->authorize('delete', $seller);
         $seller->delete();
     }
 }
